+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## v0.3.5 - Schema Add and Edit 
-  * Commit: 
+  * Commit: 2026-04-01 15:16
   * Schema Insert 
   * Schema InsertRows
   * Schema UpdateFields